feat(chart): introduce history module for batch and single data retrieval
- Added a new `history` module under `src/chart` to handle historical data fetching.
- Implemented `batch.rs` for batch processing of historical data with WebSocket integration.
- Implemented `single.rs` for single symbol historical data retrieval.
- Created a `DataChannels` struct to manage communication channels for data and symbol info.
- Enhanced error handling and logging throughout the new modules.
- Updated `mod.rs` and `lib.rs` to expose the new `history` module.
