+++ src/chart/mod.rs
@@ -2,10 +2,9 @@ pub(crate) mod options;
 pub mod study;
 pub(crate) mod utils;
 
-mod data;
+pub mod history;
 mod models;
 
-pub use data::*;
 pub use models::*;
 pub use options::ChartOptions;
 pub use options::StudyOptions;
